Move ShareConsumerTest and CustomQuotaCallbackTest to clients-integration-tests module
